pyproject.toml
[project]
name = "hbni-itv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.2",
    "aiosignal==1.3.2",
    "asyncpg==0.30.0",
    "attrs==25.3.0",
    "bcrypt>=4.3.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "directory-tree==1.0.0",
    "frozenlist==1.6.0",
    "google-api-core==2.24.2",
    "google-api-python-client==2.170.0",
    "google-auth==2.40.2",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "holidays>=0.74",
    "httplib2==0.22.0",
    "ics>=0.7.2",
    "idna==3.10",
    "isort==6.0.1",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "oauthlib==3.2.2",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pip-tools==7.4.1",
    "platformdirs==4.3.8",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pyproject-hooks==1.2.0",
    "python-dotenv==1.1.0",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rich==14.0.0",
    "rsa==4.9.1",
    "ruff==0.11.11",
    "setuptools==80.9.0",
    "tornado==6.5.1",
    "types-psycopg2>=2.9.21.20250516",
    "uritemplate==4.1.1",
    "urllib3==2.5.0",
    "wheel==0.45.1",
    "yarl==1.20.0",
]
